pyproject.toml
[project]
name = "moltui"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy>=1.26",
    "pillow>=10.0",
    "scikit-image>=0.22",
]

[project.scripts]
moltui = "moltui.app:run"
